requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
distlib==0.3.6
Django==4.1.5
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.2
filelock==3.9.0
install==1.3.5
phonenumbers==8.13.5
platformdirs==2.6.2
psycopg==3.1.8
psycopg2==2.9.5
sqlparse==0.4.3
typing_extensions==4.4.0
virtualenv==20.17.1
